Update ES-PRD-FDC-IMREM-E1.erb
Added "BLOCK\"  to the StartupCleanerProvider mainWorkingPath. This will prevent it from deleting the E:\TIS\PRD\FDC\IMREM-E1\Working\MorganStanley\LTA\, \CDM, \Reports, and \Corro  directories. The app treats folders on this level as temporary working folders and deletes them on restarts, along with any source files that might be in them.